Source code
The rapid growth of Android malware poses great
challenges to anti-malware systems because the sheer number
of malware samples overwhelms malware analysis systems. A
promising approach for speeding up malware analysis is to
classify malware samples into families so that the common
features in malware belonging to the same family can be exploited
for malware detection and inspection. Furthermore, the selection
of distinct malware samples in each family can drastically reduce
the number of malware to be analyzed. However, the existing
classification solutions are limited because of two reasons. First,
since the majority of Android malware are constructed by
inserting malicious components into popular apps, the malware’s
legitimate part may misguide the classification algorithms.
Second, the polymorphic variants of Android malware could
evade the detection by employing transformation attacks. In this
paper, we propose a novel approach that constructs frequent
subgraphs (fregraphs)